Avoid error handling & remove Zend-Diactoros dependency
Summary:
Removes the try/catch from dispatcher, allowing all error handling to
be handled upstream. This allows the removal of the Zend-Diactoros dependency
since there is no longer a need for a PSR7 implementation to return in the event
of an error.

All internal errors are now handled as SPL Exceptions, inheriting from both
LogicException and RuntimeException for the respective 500- and 400-level
problems.

Test Plan: Added unit tests, which pass

Differential Revision: https://phabricator.ericstern.com/D60