Move the error handling into a class method on Endpoints
Summary:
In the common case, there may be a couple of different output formats depending on the page being served (JSON, HTML, etc). Having a single global error handler in the front controller is inappropriate for this, since it becomes very difficult to serve a response that fulfills the requested content type.

This adds a required exception handler to each individual endpoint so that the self-contained object can return an appropriately-formatted error. It's intended (though not required) that implementors do so via a Trait (or class inheritance), so that an application may have, e.g., `HtmlErrorHandlerTrait` and `JsonErrorHandlerTrait`

There are two situations that aren't handled well, yet: 404 and if the error handler itself has a TypeError. The latter should be detectable through static analysis and unit tests (the included `EndpointTestTrait` has been updated appropriately); the former will be addressed with added functionality soon.

Test Plan: Unit tests have been added

Differential Revision: https://phabricator.ericstern.com/D73